Split parser out into separate module
